// src/app/interfaces/paciente.ts
export interface Paciente {
    id: string;
    nombre: string;
    apellido: string;
    edad: number;
    telefono: number;
    direccion: string;
    estadoCivil: string;
    profesion: string;
    estrato: number;
    cargo: string;
    horasTrabajadasDia: number;
    cantAlimentosDia: number;
    ultimoNivelEstudio: string;
    sumatoriaEncuesta: number;
}